agent_backend/requirements.txt
requests==2.32.4
google-adk==1.6.1
pydantic==2.11.3
python-dotenv==1.1.0
google-cloud-aiplatform[adk,agent_engines]>=1.42.1
absl-py==2.1.0
cloudpickle==3.0.0
fastapi
uvicorn[standard]